refactor: standardization of model options
