+++ apps/web/src/services/dashboards/config.ts
@@ -1,5 +1,7 @@
 import type { TranslateResult } from 'vue-i18n';
 
+import type { EnumModelOptions, SearchResourceModelOptions, ReferenceResourceModelOptions } from '@/models/widget';
+
 import type { DashboardLayoutWidgetInfo } from '@/services/dashboards/widgets/_configs/config';
 
 
@@ -91,24 +93,8 @@ export type VariableSelectionType = typeof VARIABLE_SELECTION_TYPES[number];
 export const VARIABLE_TYPES = ['MANAGED', 'CUSTOM'] as const;
 export type VariableType = typeof VARIABLE_TYPES[number];
 
-
-export interface EnumOptions {
-    type: 'ENUM';
-    values: { key: string; label: string; }[];
-}
-export interface SearchResourceOptions {
-    type: 'SEARCH_RESOURCE';
-    resource_type: string;
-    resource_key?: string;
-    default_path?: string|number;
-}
-export interface ReferenceResourceOptions {
-    type: 'REFERENCE_RESOURCE',
-    reference_key: string;
-}
-
 type LegacyOptions = string[];
-type VariableOptions = EnumOptions|SearchResourceOptions|ReferenceResourceOptions|LegacyOptions;
+type VariableOptions = EnumModelOptions|SearchResourceModelOptions|ReferenceResourceModelOptions|LegacyOptions;
 
 // variables schema
 export interface DashboardVariableSchemaProperty {
